address Silvio's comments
